Adjust the ECS deploy to accommodate beta
Without this change, the do-ecs-deployment script does not have the
flexibility to ship code to the beta environment.  This is due to beta
operating under different naming patterns that were used previously.

As such, we adjust the case for non-master deployments to detect the
name of the cluster.  In the event of a cluster that begins with the
string 'beta', we use the name of the repo as the service name, prefixed
with 'beta-'.

This change allows us to modify the ECS_PREPROD_CLUSTER variable in the
circle.yml of the service to be deployed in the beta environment, and
use a name compatible with the environment.

In the case of the 'more-reviews-service', the service name to push was
'app19' and with this change would be 'beta-more-reviews-service'.